+++ R/plot_grid.R
@@ -1,0 +1,22 @@
+#' Plot the grid of a tuned tskrr model
+#'
+#' With this function you can visualize the grid search for optimal
+#' lambdas from a \code{\link[xnet:tskrrTune-class]{tskrrTune}} object.
+#' In the case of a heterogenous network, this function will plot a
+#' contour plot on a grid, based on the functions \code{\link{image}}
+#' and \code{\link{contour}}
+#'
+#' @param x an object that inherits from
+#' \code{\link[xnet:tskrrTune-class]{tskrrTune}}
+#'
+#' @param ... arguments passed to image
+#'
+#' @return \code{NULL} invisibly
+#'
+#' @rdname plot_grid
+#' @export
+plot_grid <- function(x, ...){
+
+
+
+}